+++ modules/csv/src/csv-loader.js
@@ -15,30 +15,33 @@ const CSVLoader = {
   extensions: ['csv'],
   mimeType: 'text/csv',
   category: 'table',
-  parse: async (arrayBuffer, options) =>
-    parseCSVSync(new TextDecoder().decode(arrayBuffer), options),
-  parseTextSync: parseCSVSync,
+  parse: async (arrayBuffer, options) => parseCSV(new TextDecoder().decode(arrayBuffer), options),
+  parseText: parseCSV,
   parseInBatches: parseCSVInBatches,
   testText: null,
   options: {
     csv: {
       TableBatch: RowTableBatch,
-      batchSize: 10
+      batchSize: 10,
+      header: 'auto',
+      rowFormat: 'auto'
     }
   }
 };
 
 export default CSVLoader;
 
-function parseCSVSync(csvText, options) {
+async function parseCSV(csvText, options) {
   // Apps can call the parse method directly, we so apply default options here
   options = {...CSVLoader.options, ...options};
   options.csv = {...CSVLoader.options.csv, ...options.csv};
 
+  const header = await hasHeader(csvText, options);
+
   const config = {
-    header: hasHeader(csvText, options),
     dynamicTyping: true, // Convert numbers and boolean values in rows from strings
     ...options.csv,
+    header,
     download: false, // We handle loading, no need for papaparse to do it for us
     error: e => {
       throw new Error(e);
@@ -60,6 +63,8 @@ function parseCSVInBatches(asyncIterator, options) {
 
   const asyncQueue = new AsyncQueue();
 
+  const convertToObject = options.csv.rowFormat === 'object';
+
   let isFirstRow = true;
   let headerRow = null;
   let tableBatchBuilder = null;
@@ -69,7 +74,6 @@ function parseCSVInBatches(asyncIterator, options) {
     download: false, // We handle loading, no need for papaparse to do it for us
     dynamicTyping: true, // Convert numbers and boolean values in rows from strings
     header: false, // Unfortunately, header detection is not automatic and does not infer types
-
     // chunk(results, parser) {
     //   // TODO batch before adding to queue.
     //   console.log('Chunk:', results, parser);
@@ -79,10 +83,12 @@ function parseCSVInBatches(asyncIterator, options) {
     // step is called on every row
     step(results, parser) {
       const row = results.data;
+      const bytesUsed = results.meta.cursor;
 
       // Check if we need to save a header row
       if (isFirstRow && !headerRow) {
-        const header = options.header === undefined ? isHeaderRow(row) : options.header;
+        // Auto detects or can be forced with options.csv.header
+        const header = isHeaderRow(row, options);
         if (header) {
           headerRow = row;
           return;
@@ -97,19 +103,21 @@ function parseCSVInBatches(asyncIterator, options) {
 
       // Add the row
       tableBatchBuilder =
-        tableBatchBuilder || new TableBatchBuilder(TableBatchType, schema, batchSize);
+        tableBatchBuilder ||
+        new TableBatchBuilder(TableBatchType, schema, {batchSize, convertToObject});
 
       tableBatchBuilder.addRow(row);
       // If a batch has been completed, emit it
       if (tableBatchBuilder.isFull()) {
-        asyncQueue.enqueue(tableBatchBuilder.getNormalizedBatch());
+        asyncQueue.enqueue(tableBatchBuilder.getBatch({bytesUsed}));
       }
     },
 
     // complete is called when all rows have been read
     complete(results, file) {
+      const bytesUsed = results.meta.cursor;
       // Ensure any final (partial) batch gets emitted
-      const batch = tableBatchBuilder.getNormalizedBatch();
+      const batch = tableBatchBuilder.getBatch({bytesUsed});
       if (batch) {
         asyncQueue.enqueue(batch);
       }
@@ -124,27 +132,34 @@ function parseCSVInBatches(asyncIterator, options) {
   return asyncQueue;
 }
 
-function isHeaderRow(row) {
+function isHeaderRow(row, options) {
+  if (options && options.csv.header !== 'auto') {
+    return Boolean(options.csv.header);
+  }
+
   return row.every(value => typeof value === 'string');
 }
 
-function hasHeader(csvText, options) {
-  if ('header' in options) {
-    return options.header;
+async function hasHeader(csvText, options) {
+  if (options.csv.header !== 'auto') {
+    return Boolean(options.csv.header);
   }
 
-  let header = false;
-  Papa.parse(csvText, {
-    download: false,
-    dynamicTyping: true,
-    step: (results, parser) => {
-      const row = results.data;
-      header = isHeaderRow(row);
-      parser.abort();
-    }
+  return await new Promise((resolve, reject) => {
+    Papa.parse(csvText, {
+      download: false,
+      dynamicTyping: true,
+      step: (results, parser) => {
+        parser.abort();
+        const row = results.data;
+        // Test the row
+        resolve(isHeaderRow(row));
+      },
+      error: e => {
+        reject(new Error(e));
+      }
+    });
   });
-
-  return header;
 }
 
 function deduceSchema(row, headerRow) {