+++ modules/tables/src/lib/table/columnar-table-batch.d.ts
@@ -1,0 +1,15 @@
+import {Schema, Batch, IBatchBuilder} from './table-types';
+
+export default class ColumnarTableBatch  implements IBatchBuilder{
+  constructor(schema: Schema, options?: object);
+
+  addRow(row: any[] | object): void;
+
+  chunkComplete(chunk: ArrayBuffer | string): void;
+
+  isFull(): boolean;
+
+  hasBatch(): boolean;
+
+  getBatch(): Batch | null;
+}