+++ docs/whats-new.md
@@ -32,8 +32,12 @@ Worker support for the `WKTLoader`, designed to support future binary data impro
 
 **@loaders.gl/json**
 
-- `parseInBatches` now returns a `batch.bytesRead` field to enable progress bars.
+- `parseInBatches` now returns a `batch.bytesUsed` field to enable progress bars.
 - `.geojson` is now parsed by a new experimental `GeoJSONLoader` (exported with an underscore as `_GeoJSONLoader`), designed to support future binary data improvements.
+
+**@loaders.gl/csv**
+
+- Header auto-detection via `options.csv.header: 'auto'`.
 
 **@loaders.gl/arrow**